requirements.txt
Flask==2.1.0
python-dotenv==0.19.2
requests==2.27.1
transformers